Added step to skip alredy-downloaded HTML files.

+++ scripts/primate_jobs_scraper.sh
@@ -5,7 +5,9 @@ cd data/primate_jobs/
 PREFIX=http://pin.primate.wisc.edu/jobs/listings/
 
 for i in $( seq 1 9999 ); do
-    echo "--- Processing index $i --------------------------------"
-    wget ${PREFIX}${i} -O ${i}.html
-    sleep 3
+	if [ ! -s ${i}.html ]; then
+	    echo "--- Processing index $i --------------------------------"
+    	wget ${PREFIX}${i} -O ${i}.html
+    	sleep 3
+    fi
 done
